Adding stylistic Changes
In order to make the app more professional
and attractive to folks looking for an organizational
app certain styling changes are being made.

In this update you will see the beginning of a footer,
external links to provide information to the dev and users,
and support for the users.

The issue queue is now linked for users to suggest changes.